Automatically backed up by Learn

+++ lib/intro_to_simple_array_manipulations.rb
@@ -44,5 +44,5 @@ def using_delete(array, string)
 end
 
 def using_delete_at(array, Integer)
-  array.delete(Integer(array))
+  array.delete(Integer)
 end
